Rename config to producer
